refactor: sidebar folders
