+++ packages/react-ui/src/app/routes/platform/security/sso/allowed-domain.tsx
