+++ packages/react-ui/src/app/routes/platform/security/sso/oauth2-dialog.tsx
