+++ packages/react-ui/src/app/routes/platform/setup/ai-providers/ai-provider-card.tsx
