+++ packages/react-ui/src/app/routes/platform/setup/ai-providers/index.tsx
@@ -8,8 +8,8 @@ import { aiProviderApi } from '@/features/platform-admin-panel/lib/ai-provider-a
 import { AI_PROVIDERS } from '@activepieces/pieces-common';
 import { PlatformRole } from '@activepieces/shared';
 
-import { authenticationSession } from '../../../../lib/authentication-session';
-import LockedFeatureGuard from '../../../components/locked-feature-guard';
+import { authenticationSession } from '../../../../../lib/authentication-session';
+import LockedFeatureGuard from '../../../../components/locked-feature-guard';
 
 import { AIProviderCard } from './ai-provider-card';
