+++ packages/react-ui/src/app/routes/platform/setup/branding/appearance-section.tsx
