+++ packages/react-ui/src/app/routes/platform/setup/branding/smtp-section.tsx
