+++ packages/react-ui/src/app/routes/platform/setup/license-key/activate-license-dialog.tsx
