+++ packages/react-ui/src/app/routes/platform/setup/license-key/index.tsx
