+++ packages/react-ui/src/app/routes/platform/setup/templates/upsert-template-dialog.tsx
@@ -27,7 +27,7 @@ import {
   TemplateType,
 } from '@activepieces/shared';
 
-import { Textarea } from '../../../../components/ui/textarea';
+import { Textarea } from '../../../../../components/ui/textarea';
 
 const UpsertFlowTemplateSchema = Type.Object({
   displayName: Type.String({
